// client/Player.ts
export class Player {
	title: String;
	id: Number;
	first_name: String;
    second_name: String;
    squad_number: Number;
    now_cost: Number;
    total_points: Number;
    minutes: Number;
    goals_scored: Number;
    assists: Number;
    clean_sheets: Number;
    goals_conceded: Number;
    own_goals: Number;
    penalties_saved: Number;
    penalties_missed: Number;
    yellow_cards: Number;
    red_cards: Number;
    saves: Number;
    bonus: Number;
}